+++ libs/data-mapper-v2/src/models/Schema.ts
@@ -1,0 +1,16 @@
+import type { SchemaType } from '@microsoft/logic-apps-shared';
+
+export const UploadSchemaTypes = {
+  UploadNew: 'upload-new',
+  SelectFrom: 'select-from',
+} as const;
+export type UploadSchemaTypes = (typeof UploadSchemaTypes)[keyof typeof UploadSchemaTypes];
+
+export interface FileWithVsCodePath extends File {
+  path?: string;
+}
+export interface SchemaFile {
+  name: string;
+  path: string;
+  type: SchemaType;
+}
